segfault

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: amuriel <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/29 19:28:17 by amuriel           #+#    #+#             */
-/*   Updated: 2021/02/02 11:29:24 by amuriel          ###   ########.fr       */
+/*   Updated: 2021/02/24 11:51:47 by amuriel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	while (s[i] != '\0')
